refactor(game): move test data objects into testdata package
