+++ game/src/test/kotlin/com/github/bedrockecs/game/zimpl/db/world/BlockRegistryImplBlockTypesTest.kt
@@ -1,7 +1,7 @@
 package com.github.bedrockecs.game.zimpl.db.world
 
 import com.github.bedrockecs.game.db.world.BlockRegistry
-import com.github.bedrockecs.game.zimpl.common.TestAirBlockType
+import com.github.bedrockecs.game.zimpl.common.testdata.TestAirBlockType
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
